+++ NEWS.md
@@ -6,6 +6,8 @@
 
 - Visualreport: Now able to handle recordings with no valid days argument visualreport_without_invalid = TRUE, fixes bug #1016
 
+- Facilitate handling of Sensewear xls file format with externally derived epoch data #974
+
 # CHANGES IN GGIR VERSION 3.0-3
 
 - Part 2: Fix bug where data_quality_report.csv contained incorrect filehealth[...] values whenever some of these values were supposed to be blank #1003
@@ -33,6 +35,7 @@
 - Part 1:Enable use of temperature data in adhoc-csv format throughout GGIR #991
 
 - Part 1: Fix bug that caused a mismatch between IDs and filenames in part 1 when movisens participant folders did not contain the acc.bin file #994.
+
 
 # CHANGES IN GGIR VERSION 3.0-1